refactor: move state classes to domain package and split GamePhase
Addresses PR #74 review:
- Move model/state/ to domain/model/state/
- Move GamePhase enum to domain/enums/
- Split GamePhase.kt: enum in domain/enums, toDisplayText extension in domain/model/state/GamePhaseDisplay.kt
