+++ app/src/main/java/com/machikoro/client/domain/enums/GamePhase.kt
@@ -1,0 +1,9 @@
+package com.machikoro.client.domain.enums
+
+enum class GamePhase {
+    NONE,
+    ROLL_DICE,
+    RESOLVE_EFFECTS,
+    BUY_OR_BUILD,
+    END_TURN
+}
