+++ app/src/main/java/com/machikoro/client/network/websocket/WebSocketClient.kt
@@ -1,6 +1,6 @@
 package com.machikoro.client.network.websocket
 
-import com.machikoro.client.model.state.ConnectionStatus
+import com.machikoro.client.domain.model.state.ConnectionStatus
 import kotlinx.coroutines.flow.StateFlow
 
 interface WebSocketClient {
