+++ app/src/test/java/com/machikoro/client/network/websocket/OkHttpWebSocketClientTest.kt
@@ -1,6 +1,6 @@
 package com.machikoro.client.network.websocket
 
-import com.machikoro.client.model.state.ConnectionStatus
+import com.machikoro.client.domain.model.state.ConnectionStatus
 import java.io.IOException
 import okhttp3.Protocol
 import okhttp3.Request
